Fix config and clear results
Earlier tests used the commit reference functionality of asv to to rebuild the repo for each test, but since the build_command was null (due to the multiple sub-repos and multi-step build procedure), the tests were all just using the default wheel that was included as a project dependency. That will need to get working, but for now just confirm that different wheels yield expected (different) results across the dependency matrix. Add different python versions just for curiosity's sake.